save function working
