+++ pages/index.js
@@ -1,16 +1,9 @@
-import Editor from "./editor";
-
-function saveText(title = "Unnamed Document", text) {
-  localStorage.setItem(`${title}`, `${text}`);
-}
-function loadText(title) {
-  localStorage.getItem(`${title}`);
-}
+import MyApp from "./_app";
 
 export default function Home() {
   return (
     <div>
-      <Editor saveText={saveText} loadText={loadText} />
+      <MyApp />
     </div>
   );
 }
